+++ CHANGELOG.rst
@@ -39,6 +39,8 @@ Changes:
 - Overloaded the ``bind``, ``unbind``, ``try_unbind`` and ``new`` methods in the ``FilteringBoundLogger`` `Protocol <https://docs.python.org/3/library/typing.html#typing.Protocol>`_.
   This makes it easier to use objects of type ``FilteringBoundLogger`` in a typed context.
   `#392 <https://github.com/hynek/structlog/pull/392>`_
+- ``structlog.processors.StackInfoRenderer`` now has an *additional_ignores* parameter that allows you to filter out your own logging layer.
+  `#396 <https://github.com/hynek/structlog/issues/396>`_
 
 
 ----